Small cosmetic refactor
